+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changes in this release:
 - Added export to server button (#8)
 - Use separate virtual env for the compiler (#9)
 - Added symbol search support (Ctrl+T) (#34)
+- Extended language server support on Windows (#26)
 
 # v 0.0.1
 - Initial release